+++ ChangeLog.md
@@ -5,6 +5,10 @@ Rest client change log
 
 ## 2.1.0 / 2021-05-30
 
+* Added `webservices.rest.Result::link()` method to access links with a
+  given `rel` attribute, see for example how GitHub implements pagination:
+  https://docs.github.com/en/rest/overview/resources-in-the-rest-api#pagination
+  (@thekid)
 * Merged PR #14: Fluent result interface. This pull request adds a new
   `RestResponse::result()` method which returns `Result` instances with
   high-level access to responses including handling of unexpected status